Fix dangerous casts
On Windows 64-bit, the size of long is 32-bit, thus any pointer to long cast is
not valid. In any case, one should use intptr_t and ptrdiff_t when casting
to integers. This commit attempts to fix all instances reported by GCC.
When relevant, I replaced code by the macros PTR_ADD, ALIGN_UP from system.h

+++ apps/plugins/zxbox/snapshot.c
@@ -362,7 +362,7 @@ static void read_compressed_data(SNFILE *fp, byte *start, unsigned size,
     GET_DATA(ch);
     if(p + times > end) {
       put_msg("Warning: Repeat parameter too large in snapshot");
-      times = (int) ((long) end - (long) p);
+      times = (int) ((intptr_t) end - (intptr_t) p);
     }
     for(j = 0; j < times; j++) *p++ = ch;
       }
