+++ lib/rbcodec/codecs/liba52/bitstream.c
@@ -35,7 +35,7 @@ void a52_bitstream_set_ptr (a52_state_t * state, uint8_t * buf)
 {
     int align;
 
-    align = (long)buf & 3;
+    align = (intptr_t)buf & 3;
     state->buffer_start = (uint32_t *) (buf - align);
     state->bits_left = 0;
     state->current_word = 0;
